fix(usage): split management user filter expand helper
Keep list.go under the 1200-line structure gate after expanding
account-scoped API key filters for management logs.
